Updated comments in API routes
