+++ src/app/api/v1/events/[code]/matches/[name]/route.ts
@@ -2,8 +2,7 @@ import { NextResponse } from "next/server";
 import prisma from "@/lib/prisma";
 
 //mainDataSlice/getActiveTeamNumberAsync
-//Gets the teams and scouters currently set to each station during the specified match.
-//The above function is responsible for narrowing it down further to the team at a single station.
+//Gets all teams and scouters currently set to each station during the specified match.
 export async function GET(
   req: Request,
   { params }: { params: { code: string; name: string } }
