+++ src/app/api/v1/events/[code]/matches/[name]/scores/[station]/teleopCoralEvent/route.ts
@@ -1,4 +1,6 @@
 //Creates a teleopCoralScoringEvent.
+//UPDATE CYCLE: Please ensure this function's update values match the schema it's based off of.
+
 import { NextResponse } from "next/server";
 import prisma from "@/lib/prisma";
