+++ src/app/api/v1/events/[code]/statistics/all/csv/route.ts
@@ -1,20 +1,13 @@
 import { NextResponse } from "next/server";
 import prisma from "@/lib/prisma";
 import {
-  Event,
-  IncapSegment,
-  TeamScore,
-  AutoCoralScoringEvent,
-  AutoAlgaeScoringEvent,
-  TeleopCoralScoringEvent,
-  TeleopAlgaeScoringEvent,
   CoralScoringLevel,
   AlgaeScoringLocation,
 } from "@prisma/client";
-import { Ranking } from "@/lib/enums";
 
-//see DataControls for implementation.
-//If you just want raw spreadsheet info about the teams and matches in the event, here you go.
+//see the "CSV" button in DataControls for implementation.
+//Creates an spreadsheet containing data about the specified event.
+//UPDATE CYCLE: Most of the necessary edits also apply to the JSON button.
 export async function GET(
   req: Request,
   { params }: { params: { code: string } }
@@ -35,6 +28,7 @@ export async function GET(
             blue1Team: true,
             blue2Team: true,
             blue3Team: true,
+            //UPDATE CYCLE: Ensure all scoring events are listed in each of the _TeamScore objects.
             red1TeamScore: {
               include: {
                 team: true,
@@ -109,6 +103,7 @@ export async function GET(
     let matches = event.matches;
 
     const teamScoresWithDetails = matches
+      //Turns all the match data into an array of submitted teamScores.
       .flatMap((match) => [
         ...(match.red1TeamScore
           ? [
@@ -172,6 +167,8 @@ export async function GET(
           : []),
       ])
       .filter((teamScore) => teamScore.submitted)
+      //Calculates extra data about each teamScore
+      //UPDATE CYCLE: Ensure all scoring locations for all scoring events are calculated here, including dropped pieces.
       .map((teamScore) => ({
         ...teamScore,
         teamNumber: teamScore.teamNumber,
@@ -216,6 +213,7 @@ export async function GET(
         teleopCoralDropped: teamScore.teleopCoralScoringEvents.filter((event) => event.dropped).length,
         teleopAlgaeDropped: teamScore.teleopAlgaeScoringEvents.filter((event) => event.dropped).length,
       }))
+      //UPDATE CYCLE: Ensure all scoring events are listed here.
       .map((teamScore) => {
         const {
           team,
@@ -237,6 +235,8 @@ export async function GET(
     if (teamScoresWithDetails.length == 0) {
       return NextResponse.json({ ok: false, message: "No match data." });
     }
+
+    //The stuff below is just getting this data into spreadsheet form.
 
     const headers = Object.keys((teamScoresWithDetails as Object[])[0]).join(
       ","
