+++ src/app/page.tsx
@@ -85,7 +85,7 @@ export default function Home() {
             className="mx-1 mt-2"
             size="lg"
             variant="secondary"
-            href="/data"
+            href="/viewer"
           >
             Viewer (WIP)
           </Button>
